pyproject.toml
[tool.poetry]
name = "fomo-bot"
version = "0.1.0"
description = "A Discord bot that scans pinned messages for FOMO keyword"
authors = ["Clooooode <clode@clo5de.info>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
"discord.py" = "2.3.2"
python-dotenv = "1.0.0"
audioop-lts = "^0.2.1"
google-cloud-aiplatform = "^1.38.1"
kink = "^0.8.1"

[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.9.0"
ruff = "^0.11.8"
isort = "^6.0.1"
pre-commit = "^4.2.0"

[tool.ruff]
target-version = "py313"
exclude = []

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  # "UP",  # pyupgrade
]
ignore = [
  "E501", # line too long
  "B008",  # do not perform function calls in argument defaults
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 